Polytope testing for algebraic clustering
This is reconstructing code I had on my laptop, but due to quarantine I no longer have access to those files. I use Henning Seidler's polytope code, that uses cddlib (and pycddlib) to generate vertices of a simplex. The main file here is sos_tester.py, which implements the moment constraints. This will be useful for the parametric clustering problems.

Plenty of next steps ahead. Integrate this into an optimization framework, call as a function. Use the parametric calls for f(x; theta). How to channel the data and generate coefficients. Set up the binary classification problem, nuclear norm minimization, and all the rest in cvxpy. Last step: write in mosek directly?