Validation with circle model
Semialgebraic clustering works, at least to synthesizing the SDP. It is very difficult to hit rank-1 solutions. With the circle examples, rank-1 are hit in a box of <=0.25 around optimum solution (true centers). With a box of >=0.5, rank-1 are not hit with reweighted heuristics. Need to further look at this, see if I can generate redundant constraints to help.