Reweighting the combined moment matrix
Previous projects have shown that reweighted heuristic on a decomposed matrix sometimes fails to find the rank-1 solution, while the penalizing the combined moment matrix works. In the case of intersecting circles, this is borne out. Added a combined moment matrix (sloppily), and reweighted. Also tested on circles with different radii.

Need to figure out why the quartic constraint is so important in the intersecting circles example. Perhaps I am indexing incorrectly? Why does the moment (1,1) show up when the quartic constraint is present?